Fix taxi -> taxis naming
