+++ src/taxis/taxis.controller.spec.ts
@@ -1,24 +1,24 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { TaxiController } from './taxi.controller';
-import { TaxiService } from './taxi.service';
+import { TaxisController } from './taxis.controller';
+import { TaxisService } from './taxis.service';
 import { TaxiRepository } from './taxi.repository';
 import { DataSource } from 'typeorm';
 import { dataSourceMockFactory } from '../../test/mock/dataSourceMockFactory';
 
-describe('TaxiController', () => {
-  let controller: TaxiController;
+describe('TaxisController', () => {
+  let controller: TaxisController;
 
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
-      controllers: [TaxiController],
+      controllers: [TaxisController],
       providers: [
-        TaxiService,
+        TaxisService,
         TaxiRepository,
         { provide: DataSource, useFactory: dataSourceMockFactory },
       ],
     }).compile();
 
-    controller = module.get<TaxiController>(TaxiController);
+    controller = module.get<TaxisController>(TaxisController);
   });
 
   it('should be defined', () => {
