+++ src/taxis/taxis.controller.ts
@@ -8,32 +8,32 @@ import {
   Patch,
   Post,
 } from '@nestjs/common';
-import { TaxiService } from './taxi.service';
+import { TaxisService } from './taxis.service';
 import { CreateTaxiDto } from './dto/create-taxi.dto';
 import { UpdateTaxiDto } from './dto/update-taxi.dto';
-import { Taxi } from './taxi.entity';
+import { Taxi } from './entities/taxi.entity';
 import { ResponseTaxiDto } from './dto/response-taxi.dto';
 
-@Controller('taxi')
-export class TaxiController {
-  constructor(private readonly taxiService: TaxiService) {}
+@Controller('taxis')
+export class TaxisController {
+  constructor(private readonly taxisService: TaxisService) {}
 
   // 전체 택시 정보 조회
   @Get()
   getAllTaxiInfo(): Promise<Taxi[]> {
-    return this.taxiService.getAllTaxiInfo();
+    return this.taxisService.getAllTaxiInfo();
   }
 
   // 택시 Id로 택시 정보 조회
   @Get(':id')
   getTaxiInfoById(@Param('id', ParseIntPipe) taxiId: number): Promise<Taxi> {
-    return this.taxiService.getTaxiInfoById(taxiId);
+    return this.taxisService.getTaxiInfoById(taxiId);
   }
 
   // 택시 정보 등록
   @Post()
   createTaxiInfo(@Body() taxiInfo: CreateTaxiDto): Promise<ResponseTaxiDto> {
-    return this.taxiService.createTaxiInfo(taxiInfo);
+    return this.taxisService.createTaxiInfo(taxiInfo);
   }
 
   // 택시 정보 수정
@@ -42,7 +42,7 @@ export class TaxiController {
     @Param('id', ParseIntPipe) taxiId: number,
     @Body() taxiInfo: UpdateTaxiDto,
   ): Promise<ResponseTaxiDto> {
-    return this.taxiService.updateTaxiInfoById(taxiId, taxiInfo);
+    return this.taxisService.updateTaxiInfoById(taxiId, taxiInfo);
   }
 
   // 택시 정보 삭제
@@ -50,6 +50,6 @@ export class TaxiController {
   deleteTaxiInfoById(
     @Param('id', ParseIntPipe) taxiId: number,
   ): Promise<ResponseTaxiDto> {
-    return this.taxiService.deleteTaxiInfoById(taxiId);
+    return this.taxisService.deleteTaxiInfoById(taxiId);
   }
 }
