+++ src/taxis/taxis.module.ts
@@ -1,12 +1,12 @@
 import { Module } from '@nestjs/common';
-import { TaxiController } from './taxi.controller';
-import { TaxiService } from './taxi.service';
+import { TaxisController } from './taxis.controller';
+import { TaxisService } from './taxis.service';
 import { TaxiRepository } from './taxi.repository';
 import { TypeOrmModule } from '@nestjs/typeorm';
 
 @Module({
   imports: [TypeOrmModule.forFeature([TaxiRepository])],
-  controllers: [TaxiController],
-  providers: [TaxiService, TaxiRepository],
+  controllers: [TaxisController],
+  providers: [TaxisService, TaxiRepository],
 })
-export class TaxiModule {}
+export class TaxisModule {}
